requirements.txt
groq
flask
flask-cors
requests
python-dotenv
